+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha35] - 2019-12-05
+### Add
+ - Added `PasswordHashInterface` and `MixedShaPasswordHash`.
+ - Added new tests for `MixedShaPasswordHash` object.
+### Change
+ - **[BC BREAK]** `s-token` URL parameter is changed to `s-token2` to provide better security.
+ - The image URL access token extra password property can be `PasswordHashInterface` implemented object.
+
 ## [1.0.0-alpha34] - 2019-11-15
 ### Change
  - Changed travis yaml file.